not fully functional, making a survey works and adding questions as well but questions dont get updated dynamically

+++ src/main/java/org/example/repositories/Range_questionRepository.java
@@ -3,6 +3,6 @@ package org.example.repositories;
 import org.example.models.Range_question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Range_questionRepository extends JpaRepository<Range_question, Integer> {
+public interface Range_questionRepository extends JpaRepository<Range_question, Long> {
     // Add custom queries if needed
 }
